cmc/requirements.txt
Django==1.9.6
Pillow==3.2.0
Unidecode==0.04.19
Werkzeug==0.11.10
algoliasearch==1.8.0
algoliasearch-django==1.2.4
boto==2.43.0
cffi==1.7.0

cryptography==1.4
dj-database-url==0.4.1
django-autoslug==1.9.3
django-cities-light==3.3.0
django-debug-toolbar==1.4
django-extensions==1.7.1

django-storages==1.5.1
enum34==1.1.6
gunicorn==19.6.0
httplib2==0.9.2
html5lib==1.0b8
idna==2.1
ipaddress==1.0.16
lxml==3.6.4
ndg-httpsclient==0.4.0
newrelic==2.64.0.48
nose==1.3.7

progressbar2==3.9.3
psycopg2==2.6.1
pyOpenSSL==16.0.0
pyasn1==0.1.9
pycparser==2.14
python-http-client==2.2.1
python-utils==2.0.0
pytz==2016.6.1
requests==2.10.0
sendgrid==2.2.1
sendgrid-django==2.0.0
simplejson==3.8.2
six==1.10.0
smtpapi==0.3.1
sqlparse==0.1.19
unicodecsv==0.14.1
whitenoise==3.0
wsgiref==0.1.2
python-slugify
xhtml2pdf==0.0.6

